backend/requirements.txt
fastapi==0.104.0
uvicorn
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
psycopg2-binary==2.9.6
pydantic==1.10.13
python-dotenv==1.0.0
python-jose
passlib
bcrypt
```